requirements.txt
certifi==2019.3.9
dj-database-url==0.4.2
Django==2.1.7
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.7.2
django-extensions==2.1.6
gunicorn==19.9.0
Jinja2==2.10
MarkupSafe==1.1.1
olefile==0.46
Pillow==5.4.1
psycopg2==2.7.6.1
PyQt5-sip==4.19.15
pytz==2018.9
six==1.12.0
typing==3.6.4
whitenoise==4.1.2
